+++ scripts/client/client.py
@@ -1,9 +1,11 @@
 import logging
 import time
 from json import JSONDecodeError
+from typing import Any
 
 import httpx
 from a2wsgi import ASGIMiddleware
+from fastapi import Response
 
 from fractal_server.app.schemas.user import UserCreate
 from fractal_server.app.schemas.user import UserRead
@@ -31,7 +33,18 @@ DEFAULT_CREDENTIALS["password"] = "1234"  # nosec
 wsgi_app = ASGIMiddleware(app)
 wsgi_app.app.state.jobsV1 = []
 wsgi_app.app.state.jobsV2 = []
-wsgi_app.app.state.connection = None
+wsgi_app.app.state.fractal_ssh = None
+
+
+def response_json(res: Response) -> dict[str, Any]:
+    try:
+        return res.json()
+    except JSONDecodeError as e:
+        raise ValueError(
+            f"Error while parsing JSON body of response {res}.\n"
+            f"Original error:\n{str(e)}\n"
+            f"Reponse text\n:{res.text}"
+        )
 
 
 class FractalClient:
@@ -88,14 +101,7 @@ class FractalClient:
             return None
 
     def detail(self, res: httpx.Response):
-        try:
-            res_json = res.json()
-        except JSONDecodeError as e:
-            raise ValueError(
-                f"Error while parsing JSON body of response {res}.\n"
-                f"Original error:\n{str(e)}\n"
-                f"Reponse text\n:{res.text}"
-            )
+        res_json = response_json(res)
         if res_json.get("detail"):
             raise ValueError(f"WARNING: {res_json.get('detail')}")
 
@@ -107,7 +113,7 @@ class FractalClient:
             data=user.dict(exclude_none=True),
         )
         self.detail(res)
-        new_user_id = res.json()["id"]
+        new_user_id = response_json(res)["id"]
         # Make new user verified
         patch_user = UserUpdate(is_verified=True)
         res = self.make_request(
@@ -117,7 +123,7 @@ class FractalClient:
         )
         self.detail(res)
 
-        return UserRead(**res.json())
+        return UserRead(**response_json(res))
 
     def add_project(self, project: ProjectCreateV2):
         res = self.make_request(
@@ -126,7 +132,7 @@ class FractalClient:
             data=project.dict(),
         )
         self.detail(res)
-        return ProjectReadV2(**res.json())
+        return ProjectReadV2(**response_json(res))
 
     def add_dataset(self, project_id, dataset: DatasetCreateV2):
         res = self.make_request(
@@ -135,7 +141,7 @@ class FractalClient:
             data=dataset.dict(),
         )
         self.detail(res)
-        return DatasetReadV2(**res.json())
+        return DatasetReadV2(**response_json(res))
 
     def import_dataset(self, project_id, dataset: DatasetImportV2):
         res = self.make_request(
@@ -144,7 +150,7 @@ class FractalClient:
             data=dataset.dict(),
         )
         self.detail(res)
-        return DatasetReadV2(**res.json())
+        return DatasetReadV2(**response_json(res))
 
     def add_workflow(self, project_id, workflow: WorkflowCreateV2):
         res = self.make_request(
@@ -154,7 +160,7 @@ class FractalClient:
         )
         self.detail(res)
 
-        return WorkflowReadV2(**res.json())
+        return WorkflowReadV2(**response_json(res))
 
     def add_workflowtask(
         self,
@@ -171,7 +177,7 @@ class FractalClient:
         )
         self.detail(res)
 
-        return WorkflowTaskReadV2(**res.json())
+        return WorkflowTaskReadV2(**response_json(res))
 
     def add_working_task(self):
         task = TaskCreateV2(
@@ -186,7 +192,7 @@ class FractalClient:
             data=task.dict(exclude_none=True),
         )
         self.detail(res)
-        return TaskReadV2(**res.json())
+        return TaskReadV2(**response_json(res))
 
     def add_failing_task(self):
         task = TaskCreateV2(
@@ -200,7 +206,7 @@ class FractalClient:
             data=task.dict(exclude_none=True),
         )
         self.detail(res)
-        return TaskReadV2(**res.json())
+        return TaskReadV2(**response_json(res))
 
     def add_task(self, task: TaskCreateV2):
         res = self.make_request(
@@ -209,7 +215,7 @@ class FractalClient:
             data=task.dict(exclude_none=True),
         )
         self.detail(res)
-        return TaskReadV2(**res.json())
+        return TaskReadV2(**response_json(res))
 
     def whoami(self):
         res = self.make_request(
@@ -217,7 +223,7 @@ class FractalClient:
             method="GET",
         )
         self.detail(res)
-        return UserRead(**res.json())
+        return UserRead(**response_json(res))
 
     def patch_current_superuser(self, user: UserUpdate):
         me = self.whoami()
@@ -227,7 +233,7 @@ class FractalClient:
             data=user.dict(exclude_none=True),
         )
         self.detail(res)
-        return UserRead(**res.json())
+        return UserRead(**response_json(res))
 
     def submit_job(
         self,
@@ -246,7 +252,7 @@ class FractalClient:
         )
         self.detail(res)
 
-        return JobReadV2(**res.json())
+        return JobReadV2(**response_json(res))
 
     def wait_for_all_jobs(
         self,
@@ -267,8 +273,8 @@ class FractalClient:
                 method="GET",
             )
             if res.status_code != 200:
-                raise ValueError(f"Original error: {res.json()}")
-            job_statuses = [job["status"] for job in res.json()]
+                raise ValueError(f"Original error: {response_json(res)}")
+            job_statuses = [job["status"] for job in response_json(res)]
             if "submitted" not in job_statuses:
                 print("No submitted job left.")
                 return None